fix warnigs

+++ src/br/edu/cs/poo/ac/seguro/testes/TesteDAO.java
@@ -5,13 +5,11 @@ import java.io.File;
 import org.junit.jupiter.api.BeforeEach;
 
 import br.edu.cesarschool.next.oo.persistenciaobjetos.CadastroObjetos;
-import br.edu.cs.poo.ac.seguro.daos.SeguradoEmpresaDAO;
-import br.edu.cs.poo.ac.seguro.entidades.SeguradoEmpresa;
 
-public abstract class TesteDAO {
+public abstract class TesteDAO extends TesteAbstrato{
 	protected CadastroObjetos cadastro;
 	
-	protected abstract Class getClasse();
+	protected abstract Class<?> getClasse();
 	
 	protected TesteDAO() {
 		cadastro = new CadastroObjetos(getClasse());
@@ -19,11 +17,19 @@ public abstract class TesteDAO {
 	
 	@BeforeEach
 	public void setUp() {
-		String sep = File.separator;
-		File dir = new File("." + sep + getClasse().getSimpleName());
-		File[] files = dir.listFiles();
-		for (File file : files) {
-			file.delete();
-		}
+	    String sep = File.separator;
+	    File dir = new File("." + sep + getClasse().getSimpleName());
+
+	    if (!dir.exists()) {
+	        dir.mkdirs();
+	    }
+
+	    File[] files = dir.listFiles();
+	    if (files != null) {
+	        for (File file : files) {
+	            if (file != null) file.delete();
+	        }
+	    }
 	}
+
 }
